Add support team opt-ins.
The action now supports a TEAMS environment variable, which should be
formatted as a space-delimited list of team slugs, where each identified
team belongs to org that is assumed to be the owner of the repository
against which the pull request was opened (the pull request base).

When that environment variable is present, we fetch the members of the
listed teams, and skip the workflow if the PR author is not a member
of one of those teams (and is also not specified as an individual author).